local-api/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-inputs
waitress
flask_cors